creacion de componenete para almacenar 2 ordenes en una solicitud, tambien se creo modal de detalle y se modifico el modal de ver detalle de provedor con la nueva informacion y por ultimo se modifico las tabla de provedores para especificar si es una orden doble o unica

+++ database/migrations/2024_09_30_064455_create_requests_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('client_address')->nullable();
             $table->string('client_phone')->nullable();
             $table->string('container_type')->nullable();
-            $table->integer('order_weight')->nullable();
+            $table->decimal('order_weight',60,17)->nullable();
             $table->date('date_quotation')->nullable();
             $table->text('comment')->nullable();
             $table->boolean('status')->default(0);
