+++ database/migrations/2024_10_01_034521_add_gross_weight_columns_to_requests_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('requests', function (Blueprint $table) {
-            $table->integer('gross_weight')->after('order_weight')->nullable();
+            $table->decimal('gross_weight',60,17)->after('order_weight')->nullable();
         });
     }
